feat: permitir inserção manual de valores no app

+++ src/app.py
@@ -65,9 +65,9 @@ else:
 
         # Input numérico
         if info.get("type") == "numeric":
-            min_val = info.get("min", 0.0)
-            max_val = info.get("max", 100.0)
-            default = info.get("mean", (min_val + max_val)/2)
+            min_val = float(info.get("min", 0.0))
+            max_val = float(info.get("max", 100.0))
+            default = float(info.get("mean", (min_val + max_val)/2))
             input_data[feature] = st.number_input(
                 f"{feature}",
                 min_value=min_val,
